feat: create useTecnicoChamados hook and extract filter and fetch logic
