+++ hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './useAuth';
 export * from './useChamados';
+export * from './useTecnicoChamados';
 export * from './useToast';
 export * from './useUser';
